==> notebooks/01-get_data.ipynb <==
import pandas as pd
import numpy as np
import io
import sys
import os
import urllib.request
from tqdm import tqdm
from os import listdir
from PIL import Image
import glob

CSV_FILE = os.path.abspath('./csv/MovieGenre.csv')

df = pd.read_csv(CSV_FILE,encoding='ISO-8859-1')
df.head()

df.shape

check_nan_in_df = df.isnull().values.any()
print (check_nan_in_df)

#Remove NaN valued entries
df= df.dropna(how='any')
df.shape

#Method 1 to download images
IMAGES_PATH = os.path.join(os.path.abspath('../data/'), 'IMDB_IMAGES')
if (os.path.exists(IMAGES_PATH) == False):
    os.mkdir(IMAGES_PATH)

not_found = []
for index, row in tqdm(df.iterrows()):
    
    url = row['Poster']
    imdb_id = row['imdbId']
    
    file_path = os.path.join(IMAGES_PATH,  str(imdb_id) + ".jpg")
    
    try:
        response = urllib.request.urlopen(url)
        data = response.read()
        file = open(file_path, 'wb')
        file.write(bytearray(data))
        file.close()
    except:
        not_found.append(imdb_id)
        
print("Number of posters not found = ", len(not_found))
print("Following IMDB_ID posters were not found::", not_found)

IMAGES_PATH = os.path.join(os.path.abspath('../data/'), 'IMDB_IMAGES')
if (os.path.exists(IMAGES_PATH) == False):
    os.mkdir(IMAGES_PATH)

not_found = []
for index, row in tqdm(df.iterrows()):
    imdb_id = row['imdbId']
    file_path = os.path.join(IMAGES_PATH, str(imdb_id) + ".jpg")
    
    if not os.path.exists(file_path):
        not_found.append(imdb_id)

print("Number of posters not found =", len(not_found))
print("Following IMDB_ID posters were not found:", not_found)

# Remove from dataframe those whose posters were not found
df_2 = df[~df['imdbId'].isin(not_found)]
df_2.shape

# Check which downloaded images are corrupt and remove them 
IMAGES_PATH = os.path.abspath('../data/IMDB_IMAGES/')

bad_images = []
for file in glob.glob(os.path.join(IMAGES_PATH, "*.jpg")):
    try:
        img = Image.open(file) # open image file
        img.verify() # verify its an image
    except (IOError, SyntaxError) as e:
        print('Bad file:', file) 

        bad_images.append(file)
        
print("Number of corrupt files:", len(bad_images))

# No corrupt files found so not performing any further data-cleaning in terms of unfound/bad data

df_2.columns
df_2.head()

# Keep only IMDB_ID, Genres, Title

df_3 = df_2[['imdbId','Title','Genre']]
print(df_3.shape)
df_3.head()

#Checking if all the imdb_id listed here actually have its poster image

image_list = []
for file in glob.glob(os.path.join(IMAGES_PATH, "*.jpg")):
    image_list.append(file)
    
print("Number of files found:", len(image_list))

# Mis-match still exists, so we further perform data-cleaning
# This is performed in the next python script: "Clean_data"
# We export the df_3 as csv, to be used with images downloaded in the Posters folder

df_3.to_csv(os.path.join(os.path.abspath('./csv/'), 'MovieGenre_cleaned.csv'), index = None)
